fix(runtime): dom functions in circular dependency
