+++ packages/runtime/dom/index.ts
@@ -1,3 +1,2 @@
 export * from "./attrs";
 export * from "./dom";
-export * from "./elements";
